+++ lib/functions/dts-process.js
@@ -6,6 +6,7 @@ const Joi = require('joi')
 const axios = require('axios')
 const { HTTP_NOT_FOUND } = require('../constants')
 const deleteStation = require('../helpers/imtd-api').deleteStation
+const getRloiIds = require('../helpers/imtd-api').getRloiIds
 
 async function insertStation (stationDataArray) {
   try {
@@ -78,29 +79,15 @@ async function validateStationData (stationDataArray) {
   }
 }
 
-async function getRloiIds ({ limit, offset } = {}) {
-  try {
-    logger.info(`Retrieving up to ${limit} rloi_ids with an offset of ${offset}`)
-    const result = await pg('rivers_mview')
-      .distinct('rloi_id')
-      .whereNotNull('rloi_id')
-      .orderBy('rloi_id', 'asc')
-      .limit(limit)
-      .offset(offset)
-    logger.info(`Retrieved ${result.length} rloi_ids`)
-    return result
-  } catch (error) {
-    throw Error(`Could not get list of id's from database (Error: ${error.message})`)
-  }
-}
-
 async function handler ({ offset = 0 } = {}) {
   const BATCH_SIZE = parseInt(process.env.IMTD_BATCH_SIZE || '500')
 
+  logger.info(`Retrieving up to ${BATCH_SIZE} rloi_ids with an offset of ${offset}`)
   const stations = await getRloiIds({
     offset,
     limit: BATCH_SIZE
   })
+  logger.info(`Retrieved ${stations.length} rloi_ids`)
 
   for (const station of stations) {
     await getData(station.rloi_id)
